requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-groq==0.1.8
pydantic==2.8.2
faker==25.2.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
typing-extensions==4.12.2
flask==3.0.0
werkzeug==3.0.1
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
